RSRCH_test/requirements.txt
Django==3.2.7
PyYAML==5.4.1
Pillow==8.3.2
